Move update_data_template... from main to common/template_operations
